+++ ground_station/ground_station/__main__.py
@@ -1,0 +1,4 @@
+from ground_station.main import main
+
+if __name__ == "__main__":
+    main()
