+++ ground_station/ground_station/cli.py
@@ -1,0 +1,7 @@
+"""Interactive CLI entry point."""
+
+from ground_station.code import main
+
+
+def entry():
+    main()
