+++ ground_station/pyproject.toml
@@ -1,13 +1,9 @@
-[build-system]
-requires = ["setuptools>=61.0", "wheel"]
-build-backend = "setuptools.build_meta"
-
 [project]
 name = "samwise-ground-station"
 version = "1.0.0"
-description = "LoRA ground station for Samwise CubeSat communication"
+description = "LoRa ground station for Samwise CubeSat communication"
 readme = "README.md"
-requires-python = ">=3.8"
+requires-python = ">=3.11"
 authors = [
     {name = "Stanford SSI", email = "ssi@lists.stanford.edu"}
 ]
@@ -17,7 +13,6 @@ classifiers = [
     "Intended Audience :: Science/Research",
     "Topic :: Scientific/Engineering :: Aerospace",
     "Programming Language :: Python :: 3",
-    "Programming Language :: Python :: 3.8",
     "Programming Language :: Python :: 3.11",
     "Programming Language :: Python :: 3.13",
 ]
@@ -28,24 +23,16 @@ dependencies = [
     "circuitpython-hmac",
     "adafruit-blinka",
     "pydantic>=2.0.0",
-]
-
-[project.optional-dependencies]
-server = [
     "fastapi>=0.110.0",
     "uvicorn[standard]>=0.28.0",
 ]
-test = [
+
+[dependency-groups]
+dev = [
     "pytest>=7.0.0",
     "pytest-cov>=4.0.0",
     "httpx>=0.27.0",
     "starlette",
-    "fastapi>=0.110.0",
-    "uvicorn[standard]>=0.28.0",
-]
-dev = [
-    "pytest>=7.0.0",
-    "pytest-cov>=4.0.0",
     "black>=24.0.0",
     "ruff>=0.2.0",
     "pre-commit>=3.0.0",
@@ -56,20 +43,13 @@ Homepage = "https://github.com/stanford-ssi/samwise-flight-software"
 Repository = "https://github.com/stanford-ssi/samwise-flight-software"
 Issues = "https://github.com/stanford-ssi/samwise-flight-software/issues"
 
-[project.scripts]
-samwise-gs = "cli:entry"
-
-[tool.setuptools]
-py-modules = ["cli", "code", "config", "logger", "models", "protocol", "radio_commands", "radio_initialization", "server", "state", "ui"]
-
 [tool.pytest.ini_options]
-# Test discovery
+pythonpath = ["."]
 testpaths = ["tests"]
 python_files = "test_*.py"
 python_classes = "Test*"
 python_functions = "test_*"
 
-# Output
 addopts = [
     "-v",
     "--tb=short",
@@ -77,7 +57,6 @@ addopts = [
     "--disable-warnings",
 ]
 
-# Test markers
 markers = [
     "unit: Unit tests for individual components",
     "integration: Integration tests for component interactions",
@@ -90,7 +69,7 @@ markers = [
 minversion = "7.0"
 
 [tool.coverage.run]
-source = ["."]
+source = ["ground_station"]
 omit = [
     "tests/*",
     "*/tests/*",
@@ -109,7 +88,7 @@ exclude_lines = [
 
 [tool.black]
 line-length = 100
-target-version = ['py38', 'py39', 'py310', 'py311']
+target-version = ['py311']
 exclude = '''
 /(
     \.git
@@ -132,12 +111,12 @@ exclude = [
 
 [tool.ruff.lint]
 select = ["E", "F", "W", "I", "N"]
-ignore = ["E501"]  # Line too long
+ignore = ["E501"]
 
 [tool.ruff.lint.isort]
-known-first-party = ["cli", "code", "config", "logger", "models", "protocol", "radio_commands", "radio_initialization", "server", "state", "ui"]
+known-first-party = ["ground_station"]
 
 [tool.ruff.lint.per-file-ignores]
-"__init__.py" = ["F401"]  # Unused imports in __init__ are OK
-"models.py" = ["N815", "N803"]  # Allow mixedCase for satellite data field names
-"tests/test_compatibility.py" = ["N806", "E712"]  # Allow UPPER_CASE and boolean comparisons in tests
+"__init__.py" = ["F401"]
+"ground_station/models/*.py" = ["N815", "N803"]
+"tests/test_compatibility.py" = ["N806", "E712"]
